requirements.txt
Django==1.11.3
django-cors-headers==2.1.0
djangorestframework==3.6.3
olefile==0.44
Pillow==4.2.1
PyMySQL==0.7.11
pytz==2017.2
django_nose==1.4.4
cloudconvert==1.0.0
